meson: Write command line into config-cmdline.log file
Unfortunately, current meson versions do not write the command line they
were called with in a log file. This is a useful feature for when one has
to re-run the configuration step. So, replicate this feature manually by
writing the command line to a config-cmdline.log file that resides in the
build directory.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>