Add ability to create local git repository for tarball based builds
Versioned builds typically build from release tarballs. These do not
contain a .git folder, which means that any source code modifications will
go unnoticed. This is a problem if the local installation is used for
testing and developing patches. By having the script automatically create
a local git repository, this is solved.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>